course builder bugs fixed

+++ server/controllers/Section.js
@@ -1,6 +1,6 @@
 const Section = require("../models/Section")
 const Course = require("../models/Course")
-
+const SubSection = require("../models/SubSection");
 exports.createSection = async(req,res)=>{
     try{
         const {sectionName,courseId} = req.body
@@ -40,15 +40,22 @@ exports.createSection = async(req,res)=>{
 
 exports.updateSection = async (req, res) => {
 	try {
-		const { sectionName, sectionId } = req.body;
+		const { sectionName, sectionId  , courseId} = req.body;
 		const section = await Section.findByIdAndUpdate(
 			sectionId,
 			{ sectionName },
 			{ new: true }
 		);
+        const course = await Course.findById(courseId).populate({
+            path:"courseContent",
+            populate:{
+                path:"subSection"
+            },
+        }).exec()
 		res.status(200).json({
 			success: true,
 			message: section,
+            data:course
 		});
 	} 
 	catch (error) {
@@ -61,21 +68,47 @@ exports.updateSection = async (req, res) => {
 };
 
 exports.deleteSection = async (req,res)=>{
-    try{
-        const {sectionId} = req.params
-        const updatedSection = await Section.findByIdAndDelete(sectionId)
-        return res.status(200).json({
-            sucess:true,
-            message:"Section deleted successfully",
-            updatedSection
-        })
+    try {
 
-    }
-    catch(err){
-        console.log(err)
-        return res.status(400).json({
-            success:false,
-            message:"Something went wrong while deleting the section"
-        })
-    }
+		const { sectionId, courseId }  = req.body;
+		await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+		const section = await Section.findById(sectionId);
+		console.log(sectionId, courseId);
+		if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
+
+		//delete sub section
+		await SubSection.deleteMany({_id: {$in: section.subSection}});
+
+		await Section.findByIdAndDelete(sectionId);
+
+		//find the updated course and return 
+		const course = await Course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
+
+		res.status(200).json({
+			success:true,
+			message:"Section deleted",
+			data:course
+		});
+	} catch (error) {
+		console.error("Error deleting section:", error);
+		res.status(500).json({
+			success: false,
+			message: "Internal server error",
+		});
+	}
 }
